Rewrite pathfinding APIs
This implementation aims to solve the following issues with the
pathfinding methods:

- APIs are inconsistent and limited
- algorithms should be of the box, no extra code needed
- (near-)zero runtime overhead
- non-breaking change

Currently, the customization of the pathfinding is done using
strongly-typed fluent/builder style construction. The upside of this is
that the API is self-documenting and pairs well with completion engines.
A downside is that the code is rather heavy in boilerplate code and
generics.